MGDOBR-234: split RhoasClientTest to avoid WireMockServer hanging
