+++ test-utils/src/main/java/com/redhat/service/bridge/test/rhoas/AbstractApiMockServerConfigurator.java
@@ -18,6 +18,9 @@ public abstract class AbstractApiMockServerConfigurator {
 
     private static final ObjectMapper MAPPER = new ObjectMapper();
 
+    @ConfigProperty(name = "rhoas-mock-server.delay", defaultValue = "250")
+    int delay;
+
     private final String intermediatePath;
     private final String expectedAccessToken;
 
@@ -60,10 +63,12 @@ public abstract class AbstractApiMockServerConfigurator {
                 .withStatus(status)
                 .withHeader("Content-Type", "application/json")
                 .withBody(bodyString)
-                .withFixedDelay(1000);
+                .withFixedDelay(delay);
     }
 
     protected ResponseDefinitionBuilder responseWithStatus(int status) {
-        return WireMock.aResponse().withStatus(status).withFixedDelay(1000);
+        return WireMock.aResponse()
+                .withStatus(status)
+                .withFixedDelay(delay);
     }
 }
